requirements.txt
Django>=4.0.4
django-extensions~=3.1.0
djangorestframework~=3.13.0
Markdown~=3.3.0
django-filter~=21.1
drf-spectacular~=0.22.0
django-oauth-toolkit~=2.0.0
python-jose~=3.3.0
python-dotenv~=0.20.0
pytz>=2022.0
djangorestframework-simplejwt~=5.1.0
ariadne_django~=0.2.0
ariadne~=0.15.1
requests~=2.27.0
